Python 3 compatibility. Readers and writers should expect unicode data.
