+++ .travis.yml
@@ -3,7 +3,6 @@ python:
 - 2.7
 - 3.6
 env:
-- TOXENV=django18
 - TOXENV=django111
 - TOXENV=django20
 matrix:
