+++ pylintrc
@@ -26,17 +26,28 @@
 # 1. Edit the pylintrc file in the edx-lint repo at
 #    https://github.com/edx/edx-lint/blob/master/edx_lint/files/pylintrc
 #
-# 2. Make a new version of edx_lint, which involves the usual steps of
-#    incrementing the version number, submitting and reviewing a pull
-#    request, and updating the edx-lint version reference in this repo.
+# 2. install the updated version of edx-lint (in edx-lint):
 #
-# 3. Install the newer version of edx-lint.
+#       $ pip install .
 #
-# 4. Run:
+# 3. Run (in edx-lint):
 #
+#       # uses pylintrc_tweaks from edx-lint for linting in edx-lint
+#       # NOTE: Use Python 3.x, which no longer includes comments in the output file
 #       $ edx_lint write pylintrc
 #
-# 5. This will modify the local file.  Submit a pull request to get it
+# 4. Make a new version of edx_lint, submit and review a pull request with the
+#    pylintrc update, and after merging, update the edx-lint version by
+#    creating a new tag in the repo (uses pbr).
+#
+# 5. In your local repo, install the newer version of edx-lint.
+#
+# 6. Run:
+#
+#       # uses local pylintrc_tweaks
+#       $ edx_lint write pylintrc
+#
+# 7. This will modify the local file.  Submit a pull request to get it
 #    checked in so that others will benefit.
 #
 #
@@ -53,7 +64,7 @@
 #
 # ------------------------------
 [MASTER]
-ignore = migrations
+ignore = migrations,tests
 persistent = yes
 load-plugins = caniusepython3.pylint_checker,edx_lint.pylint,pylint_django,pylint_celery
 
@@ -248,7 +259,6 @@ enable =
 	too-many-statements,
 	too-many-boolean-expressions,
 	
-	ungrouped-imports,
 	wrong-import-order,
 	wrong-import-position,
 	wildcard-import,
@@ -335,6 +345,7 @@ disable =
 	too-many-locals,
 	too-many-public-methods,
 	too-many-return-statements,
+	ungrouped-imports,
 	unichr-builtin,
 	unicode-builtin,
 	unpacking-in-except,
@@ -366,8 +377,8 @@ no-docstring-rgx = __.*__$|test_.+|setUp$|setUpClass$|tearDown$|tearDownClass$|M
 docstring-min-length = 5
 
 [FORMAT]
-max-line-length = 120
-ignore-long-lines = ^\s*(# )?<?https?://\S+>?$
+max-line-length = 140
+ignore-long-lines = ^\s*(# )?((<?https?://\S+>?)|(\.\. \w+: .*))$
 single-line-if-stmt = no
 no-space-check = trailing-comma,dict-separator
 max-module-lines = 1000
@@ -440,4 +451,4 @@ int-import-graph =
 [EXCEPTIONS]
 overgeneral-exceptions = Exception
 
-# 6221f5507b1554d42ba7e841083736b0e7ebed77
+# 756ad8fda727aefa99d1946f1249b28c2d7b943f
