+++ pylintrc_tweaks
@@ -1,4 +1,7 @@
 # pylintrc tweaks for use with edx_lint.
 [MASTER]
-ignore = migrations
+ignore = migrations,tests
 load-plugins = caniusepython3.pylint_checker,edx_lint.pylint,pylint_django,pylint_celery
+
+[FORMAT]
+max-line-length = 140
