+++ super_csv/admin.py
@@ -1,7 +1,7 @@
 """
 Django admin for CSV Operations
 """
-from __future__ import absolute_import, unicode_literals, print_function
+from __future__ import absolute_import, print_function, unicode_literals
 
 from django.contrib import admin
