+++ super_csv/exceptions.py
@@ -1,0 +1,7 @@
+"""
+Super CSV exceptions
+"""
+
+
+class ValidationError(ValueError):
+    pass
