+++ tox.ini
@@ -1,5 +1,5 @@
 [tox]
-envlist = {py27,py36}-django{18,111},py36-django20
+envlist = {py27,py36}-django111,py36-django20
 
 [doc8]
 max-line-length = 120
@@ -67,11 +67,11 @@ deps =
     -r{toxinidir}/requirements/quality.txt
 commands =
     touch tests/__init__.py
-    pylint super_csv tests test_utils manage.py setup.py
-    pylint --py3k super_csv tests test_utils manage.py setup.py
+    pylint super_csv manage.py setup.py
+    pylint --py3k super_csv manage.py setup.py
     rm tests/__init__.py
-    pycodestyle super_csv tests manage.py setup.py
-    pydocstyle super_csv tests manage.py setup.py
+    pycodestyle super_csv manage.py setup.py
+    ;pydocstyle super_csv manage.py setup.py
     isort --check-only --diff --recursive tests test_utils super_csv manage.py setup.py test_settings.py
     make selfcheck
